pyproject.toml
[project]
name = "dspy-kg"
version = "0.1.0"
description = "Knowledge graph extraction using DSPy"
authors = [
    {name = "User", email = "user@example.com"}
]
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "dspy-ai>=2.4.0",
    "pydantic>=2.0.0",
    "networkx>=3.0",
    "matplotlib>=3.7.0",
    "nltk>=3.8.0",
    "litellm[proxy]",
    "duckdb>=0.9.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/extract"]

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 88
